change search form to server action
